Cache systemd service files in `~/.cache/guix-install-action`
After `guix pull` the systemd service files no longer exist in the
current profile, therefore just restoring the Guix store and profile
symlinks is not enough to install the updated Guix on the next workflow
run.  Assume that the systemd service files did not have important
changes, and therefore can be reused (this is what would actually happen
on the initial run too, so it should not change the result, especially
if the initial Guix tarball is taken from the release instead of the
latest snapshot).